requirements.txt
# HBase数据迁移验证系统依赖库

# 核心依赖
happybase>=1.2.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0

# 数据处理
numpy>=1.21.0
openpyxl>=3.0.0

# 配置管理
PyYAML>=6.0

# 日志和监控
tqdm>=4.60.0

# 开发工具 (可选)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
